[Fix] Rename config upload component to match naming style of other components
