+++ frontend/src/routes/config/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import FileUpload from '$lib/components/File_upload.svelte';
+	import FileUpload from '$lib/components/FileUpload.svelte';
 </script>
 
 <FileUpload />
